完全版

+++ src/main/java/com/qfedu/yikitchen/service/impl/FocusServiceImpl.java
@@ -2,6 +2,7 @@ package com.qfedu.yikitchen.service.impl;
 
 import com.qfedu.yikitchen.mapper.FocusMapper;
 import com.qfedu.yikitchen.pojo.Focus;
+import com.qfedu.yikitchen.pojo.Like;
 import com.qfedu.yikitchen.service.FocusService;
 import com.qfedu.yikitchen.util.ResultUtil;
 import com.qfedu.yikitchen.vo.ResultBean;
@@ -15,19 +16,14 @@ public class FocusServiceImpl implements FocusService {
 
     @Override
     public ResultBean FocusOrNo(Focus focus) {
-        ResultBean resultBean = ResultUtil.execOp(focusMapper.FocusOrNo(focus),"查询成功");
-        return  resultBean;
-    }
-
-    @Override
-    public ResultBean addFocus(Focus focus) {
-        ResultBean resultBean = ResultUtil.execOp(focusMapper.insert(focus),"关注成功");
-        return  resultBean;
-    }
-
-    @Override
-    public ResultBean deleteFocus(Focus focus) {
+       Focus focus1=focusMapper.FocusOrNo(focus);
+    if (focus1!=null){
         ResultBean resultBean = ResultUtil.execOp(focusMapper.delete(focus),"已取消关注");
         return  resultBean;
+        }else{
+        ResultBean resultBean = ResultUtil.execOp(focusMapper.insert(focus),"关注成功");
+        return  resultBean;
+        }
+
     }
 }
