+++ src/main/java/com/qfedu/yikitchen/service/impl/LikeServiceImpl.java
@@ -1,6 +1,7 @@
 package com.qfedu.yikitchen.service.impl;
 
 import com.qfedu.yikitchen.mapper.LikeMapper;
+import com.qfedu.yikitchen.mapper.LookMapper;
 import com.qfedu.yikitchen.pojo.Like;
 import com.qfedu.yikitchen.service.LikeService;
 import com.qfedu.yikitchen.util.ResultUtil;
@@ -13,27 +14,34 @@ public class LikeServiceImpl implements LikeService {
 
     @Autowired
     private LikeMapper likeMapper;
+    @Autowired
+    private LookMapper lookMapper;
+
+
+
     @Override
     public ResultBean selectLike(Like like) {
-        ResultBean resultBean = ResultUtil.setOK("查询结果",likeMapper.likeOrNo(like));
-        return  resultBean;
+        Like like1=likeMapper.likeOrNo(like);
+
+
+        if (like1!=null){
+            ResultBean resultBean = ResultUtil.execOp(likeMapper.delete(like),"点赞取消");
+            return  resultBean;
+
+        }else {
+            ResultBean resultBean = ResultUtil.execOp(likeMapper.insert(like),"点赞成功");
+
+            return  resultBean;
+        }
+
     }
 
-    @Override
-    public ResultBean addLike(Like like) {
-        ResultBean resultBean = ResultUtil.execOp(likeMapper.insert(like),"点赞成功");
-        return  resultBean;
-    }
+
 
     @Override
-    public ResultBean deleteLike(Like like) {
-        ResultBean resultBean = ResultUtil.execOp(likeMapper.delete(like),"点赞取消");
-        return  resultBean;
-    }
+    public ResultBean LikeNum() {
 
-    @Override
-    public ResultBean LikeNum(Integer id) {
-        ResultBean resultBean =ResultUtil.setOK("查找成功",likeMapper.LikeNum(id));
+        ResultBean resultBean =ResultUtil.setOK("查找成功",likeMapper.LikeNum());
         return resultBean;
     }
 }
